backend/requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3.0
django-redis==5.4.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Async Tasks
celery==5.3.4
redis==5.0.1
flower==2.0.1

# WSGI/ASGI
gunicorn==21.2.0
uvicorn==0.23.2
whitenoise==6.6.0

# Utils
python-dotenv==1.0.0
Pillow==10.1.0
pytz==2023.3
requests==2.31.0
pyyaml==6.0.1
python-dateutil==2.8.2

# Validation
pydantic==2.4.2
email-validator==2.1.0.post1

# Monitoring
prometheus-client==0.17.1
django-prometheus==2.3.1
sentry-sdk==1.32.0

# Testing
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
faker==19.12.0

# Documentation
drf-yasg==1.21.7
markdown==3.5.1

# Security
django-ratelimit==4.1.0
django-defender==0.9.7

# Performance
django-cacheops==7.0.2
django-debug-toolbar==4.2.0

# Development
black==23.10.1
flake8==6.1.0
isort==5.12.0
mypy==1.6.1
